+++ .eslintrc.yml
@@ -2,7 +2,7 @@ parserOptions:
   sourceType: 'module'
   ecmaVersion: 2015
 
-  env:
+env:
   es6: true
   browser: true
   commonjs: true
