+++ src/view/trip-info.js
@@ -1,0 +1,8 @@
+const createTripInfoTemplate = () => {
+  return (
+    `<section class="trip-main__trip-info  trip-info">
+    </section>`
+  );
+};
+
+export {createTripInfoTemplate};
